backend/pyproject.toml
[project]
name = "llm-agent-backend"
version = "0.1.0"
description = "Backend for LLM Agent Web App Template"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "llm-agent @ {path = \"../agent\", editable = true}",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

